+++ src/fastlife/config/configurator.py
@@ -42,7 +42,7 @@ from fastlife.domain.model.template import InlineTemplate
 from fastlife.middlewares.base import AbstractMiddleware
 from fastlife.service.check_permission import check_permission
 from fastlife.service.csrf import check_csrf
-from fastlife.service.job import JobHook, JobSchedulerTrigger, Undefined, undefined
+from fastlife.service.job import JobHandler, JobSchedulerTrigger, Undefined, undefined
 from fastlife.service.registry import DefaultRegistry, TRegistry
 from fastlife.settings import Settings
 from fastlife.shared_utils.infer import is_inline_template_returned
@@ -723,7 +723,7 @@ class GenericConfigurator(Generic[TRegistry]):
 
     def register_job(
         self,
-        job: JobHook[TRegistry],
+        job: JobHandler[TRegistry],
         /,
         *,
         trigger: JobSchedulerTrigger,
@@ -738,6 +738,26 @@ class GenericConfigurator(Generic[TRegistry]):
         replace_existing: bool = False,
         **trigger_args: Any,
     ) -> None:
+        """
+        Register a scheduled job.
+
+        :param job: the job handler to run.
+        :param trigger: the way the job is triggered.
+        :param id: optional identifier for the job.
+        :param name: optional name for the job.
+        :param misfire_grace_time: seconds after the designated runtime that the job is still
+            allowed to be run (or `None` to allow the job to run no matter how late it is)
+        :param coalesce: run once instead of many times if the scheduler determines that the
+            job should be run more than once in succession
+        :param max_instances: maximum number of concurrently running instances allowed for this
+            job
+        :param next_run_time: when to first run the job, regardless of the trigger (pass
+            `None` to add the job as paused)
+        :param jobstore: alias of the job store to store the job in
+        :param executor: alias of the executor to run the job with
+        :param replace_existing: `True` to replace an existing job with the same `id`
+            (but retain the number of runs from the existing one)
+        """
         self.registry.job_scheduler.register_job(
             job,  # type: ignore
             trigger=trigger,