+++ src/fastlife/service/job.py
@@ -1,6 +1,5 @@
 """
-Base class for scheduled job.
-
+Service for scheduled jobs.
 """
 
 from collections.abc import Awaitable, Callable
@@ -16,7 +15,7 @@ from fastlife.service.registry import TRegistry
 from fastlife.shared_utils.resolver import resolve
 
 __all__ = [
-    "JobHook",
+    "JobHandler",
     "JobScheduler",
     "JobSchedulerTrigger",
     "Undefined",
@@ -27,12 +26,12 @@ JobSchedulerTriggerLiteral = Literal["interval", "cron", "date"]
 JobSchedulerTrigger: TypeAlias = JobSchedulerTriggerLiteral | BaseTrigger
 
 
-JobHook = Callable[[TRegistry], None] | Callable[[TRegistry], Awaitable[None]]
+JobHandler = Callable[[TRegistry], None] | Callable[[TRegistry], Awaitable[None]]
 
 
 class JobScheduler(Generic[TRegistry]):
     """
-    Build the registry of job using venusian.
+    Job scheduler and registry.
     """
 
     def __init__(self, registry: TRegistry) -> None:
@@ -41,7 +40,7 @@ class JobScheduler(Generic[TRegistry]):
 
     def register_job(
         self,
-        job: JobHook[TRegistry],
+        job: JobHandler[TRegistry],
         /,
         *,
         trigger: JobSchedulerTrigger,
@@ -56,6 +55,26 @@ class JobScheduler(Generic[TRegistry]):
         replace_existing: bool = False,
         **trigger_args: Any,
     ) -> None:
+        """
+        Register a scheduled job.
+
+        :param job: the job handler to run.
+        :param trigger: the way the job is triggered.
+        :param id: optional identifier for the job.
+        :param name: optional name for the job.
+        :param misfire_grace_time: seconds after the designated runtime that the job is still
+            allowed to be run (or `None` to allow the job to run no matter how late it is)
+        :param coalesce: run once instead of many times if the scheduler determines that the
+            job should be run more than once in succession
+        :param max_instances: maximum number of concurrently running instances allowed for this
+            job
+        :param next_run_time: when to first run the job, regardless of the trigger (pass
+            `None` to add the job as paused)
+        :param jobstore: alias of the job store to store the job in
+        :param executor: alias of the executor to run the job with
+        :param replace_existing: `True` to replace an existing job with the same `id`
+            (but retain the number of runs from the existing one)
+        """
         self.scheduler.add_job(  # type: ignore
             job,
             kwargs={"registry": self.registry},
@@ -73,7 +92,9 @@ class JobScheduler(Generic[TRegistry]):
         )
 
     def start(self) -> None:
+        """Start the scheduler."""
         self.scheduler.start()
 
     def shutdown(self, *, wait: bool) -> None:
+        """Shutdown the scheduler."""
         self.scheduler.shutdown(wait=wait)