+++ tests/fastlife_app/jobs/ping_me.py
@@ -1,16 +1,16 @@
 import logging
 
-from fastlife.config.jobs import sheduled_job
+from fastlife.config.jobs import scheduled_job
 from tests.fastlife_app.config import MyRegistry
 
 log = logging.getLogger(__name__)
 
 
-@sheduled_job(trigger="interval", seconds=5, max_instances=1)
+@scheduled_job(trigger="interval", seconds=5, max_instances=1)
 def ping_me(registry: MyRegistry):
     log.info("ping")
 
 
-@sheduled_job(trigger="interval", seconds=5, max_instances=1)
+@scheduled_job(trigger="interval", seconds=5, max_instances=1)
 async def pong_me(registry: MyRegistry):
     log.info("pong")